chore: remover duplicado backend/mailer.js; usar utils/mailer.js
