+++ backend/controllers/authController.js
@@ -2,7 +2,7 @@ const bcrypt = require('bcryptjs');
 const jwt = require('jsonwebtoken');
 const fs = require('fs');
 const Usuario = require('../models/Usuario');
-const { enviarCodigoVerificacao } = require('../mailer');
+const { enviarCodigoVerificacao } = require('../utils/mailer');
 const { initDB, getDBPath } = require('../db');
 const emailService = require('../services/emailService');
 const { setResetToken, getResetToken, deleteResetToken } = require('../services/userService');
